% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_calibrations.R
\name{correct_steps}
\alias{correct_steps}
\title{Undo calibration steps}
\usage{
correct_steps(x, y, steps)
}
\arguments{
\item{x}{Input x values}

\item{y}{Input y values}

\item{steps}{list of step information from find_steps or review_steps}
}
\value{
list of step x values, beginning and end y values, and step distance
}
\description{
Undo calibration steps
}
\examples{
correct_steps(1:100, c(rep(0, 30), rep(1, 70)), list(i=30, dy=1))
}
