% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample}
\alias{downsample}
\title{Downsample an Artemisia dataframe}
\usage{
downsample(data, t1, t2, dt_out_sec, fields = character())
}
\arguments{
\item{data}{Data frame to downsample, or file name containing the data}

\item{t1}{start time (UTC), either string (various YMD[HMS] formats accepted) or POSIXt type}

\item{t2}{end time (UTC), either string (various YMD[HMS] formats accepted) or POSIXt type}

\item{dt_out_sec}{sample interval for downsampled output (seconds)}

\item{fields}{vector of field names to downsample (default is all CO2, temp, RH, battery fields)}
}
\value{
Dataframe containing fields 'time' and all user-specified data fields.
}
\description{
Downsample an Artemisia dataframe
}
\details{
For each field identified by the user, this function applies an
anti-aliasing filter (sixth-order Butterworth, causal, low-pass below the
new Nyquist frequency) and downsamples the data. It attempts to preserve
missing sections of data (NaNs) in this process (as opposed to interpolating
through them, which could give inaccurate results for long sensor dropouts).
}
\examples{
## In this example, 'data' can either by the output of read_data(), or a file name with path
downsample(data, t1 = '20230922', t2 = '2023-10-12 23:59:59', dt_out_sec = 15*60)
}
