% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{fix_file_time}
\alias{fix_file_time}
\title{Correct file times}
\usage{
fix_file_time(filename, starttime = NULL, endtime = NULL)
}
\arguments{
\item{filename}{string of existing file name to process}

\item{starttime}{Optional; user-provided file start time as YYYYmmdd_HHMMSS string.}

\item{endtime}{Optional; user-provided file end time as YYYYmmdd_HHMMSS string}
}
\value{
writes output file with the normal file name based on start time and serial number
}
\description{
Fix times for data file with initial missing GPS, and write new file
}
\note{
If neither starttime nor endtime is provided, they will be calculated using GPS data found in the file. This is normally the best approach unless a file has no GPS data at all. Only one of starttime and endtime should be given; if both are provided, starttime will be ignored.
}
\examples{
medfilt_nan(data, 5)
}
