% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{medfilt_nan}
\alias{medfilt_nan}
\title{Median filter for Rtemisia data frame}
\usage{
medfilt_nan(data, n = 3)
}
\arguments{
\item{data}{Data frame from read_data()}

\item{n}{integer; width (diameter) of moving median window}
}
\value{
median-filtered data frame
}
\description{
Runs a median filter on all data columns (not time, location, etc)
}
\examples{
medfilt_nan(data, 5)
}
