% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_key}
\alias{plot_key}
\title{Plot one data type for all sensors}
\usage{
plot_key(
  x,
  key,
  xlim = NULL,
  ylim = NULL,
  interactive = TRUE,
  medfilt_radius = 18,
  ...
)
}
\arguments{
\item{x}{Data frame to plot (output of [read_data()])}

\item{key}{plot fields in data frame whose names contain key}

\item{xlim}{x-limits for plot (seconds after Jan 1 1970; default range(x$time))}

\item{ylim}{y-limits to plot (default full range of fields that are plotted)}

\item{interactive}{if TRUE, allow zooming in using a very basic GUI}

\item{medfilt_radius}{If nonzero, radius of window for median filter (helps eliminate spurious spikes in data)}

\item{...}{additional arguments to pass to plot()}
}
\value{
None
}
\description{
plot_CO2, plot_temp, and plot_RH are all wrappers for plot_key. All of them
include an optional GUI to zoom in on regions of interest (disable with
"interactive = FALSE")
}
\details{
In interactive mode, the plot functions use a very basic GUI to
navigate. Clicking twice at opposite corners of a region of interest will 
zoom to the x and y limits clicked. Clicking left once followed by a right
click will return to the original x and y limits. Finally, a single right
click will quit the GUI and return control to the command line. Note that
ctrl-c does not work immediately in interactive mode; it takes effect after
the next right click.
}
\examples{
plot_key(x, key = 'CO2', interactive = FALSE)
}
