% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plots}
\alias{plots}
\alias{plot_CO2}
\alias{plot_temp}
\alias{plot_RH}
\alias{plot_batt}
\title{Plot one data type for all sensors}
\usage{
plot_CO2(x, medfilt_radius = 30, ...)

plot_temp(x, medfilt_radius = 30, ...)

plot_RH(x, medfilt_radius = 30, ...)

plot_batt(x, medfilt_radius = 30, ...)
}
\arguments{
\item{x}{Data frame to plot (output of [read_data()])}

\item{medfilt_radius}{If nonzero, radius of window for median filter (helps eliminate spurious spikes in data)}

\item{...}{additional arguments to pass to plot_key() or plot()}
}
\value{
None
}
\description{
plot_CO2, plot_temp, and plot_RH are all wrappers for plot_key. All of them
include an optional GUI to zoom in on regions of interest (disable with
"interactive = FALSE")
}
\details{
In interactive mode, the plot functions use a very basic GUI to
navigate. Clicking twice at opposite corners of a region of interest will 
zoom to the x and y limits clicked. Clicking left once followed by a right
click will return to the original x and y limits. Finally, a single right
click will quit the GUI and return control to the command line. Note that
ctrl-c does not work immediately in interactive mode; it takes effect after
the next right click.
}
\examples{
plot_CO2()
plot_temp()
plot_RH()
plot_batt()
}
