% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{quicklook}
\alias{quicklook}
\title{Quick overview of Artemisia data}
\usage{
quicklook(x)
}
\arguments{
\item{x}{Data frame to plot (output of [read_data()])}
}
\value{
None
}
\description{
Make plots of CO2, temperature, and relative humidity for all sensors, and battery voltage/GPS
}
\examples{
quicklook()
}
\seealso{
[plot_CO2], [plot_temp], [plot_RH], [plot_key]
}
