\name{Rtemisia-package}
\alias{Rtemisia-package}
\alias{Rtemisia}
\docType{package}
\title{
\packageTitle{Rtemisia}
}
\description{
\packageDescription{Rtemisia}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Rtemisia}
\packageIndices{Rtemisia}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Rtemisia}

Maintainer: \packageMaintainer{Rtemisia}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
